model/SR_Script/pyproject.toml
[project]
name = "SR_Script"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = [
    "altgraph==0.17.4",
    "cffi==1.17.1",
    "coloredlogs==15.0.1",
    "cryptography==44.0.0",
    "filelock==3.17.0",
    "flatbuffers==25.1.24",
    "fsspec==2025.2.0",
    "humanfriendly==10.0",
    "Jinja2==3.1.5",
    "MarkupSafe==3.0.2",
    "mpmath==1.3.0",
    "networkx==3.4.2",
    "numpy==2.2.2",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-cusparselt-cu12==0.6.2",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "onnxruntime-gpu==1.20.1",
    "opencv-python==4.11.0.86",
    "packaging==24.2",
    "pillow==11.1.0",
    "protobuf==5.29.3",
    "pycparser==2.22",
    "pyinstaller==6.12.0",
    "pyinstaller-hooks-contrib==2025.1",
    "sympy==1.13.1",
    "torch==2.6.0",
    "triton==3.2.0",
    "typing-extensions==4.12.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
